[dcpmm] Add template for the dcpmm plugin
Template included for the dcpmm plugin.
All the config values are optional, and so there is no
default/main/dcpmm.yml included.
Config values are documented in README.rst

Support for dcpmm plugin is not yet included in the container used for
testing with molecule, so tests are not included.
Once the container supports dcpmm, then the plugin config can be tested
by modifying the plugin lists in tests/test.yml and
molecule/default/converge.yml.