[mdevents] Add template for the mdeventsplugin
Template included for the mdevents plugin.
All the config values are optional, and so there is no
default/main/mdevents.yml included.
Config values are documented in README.rst

Support for mdevents plugin is not yet included in the container used for
testing with molecule, so tests are not included.
Once the container supports mdevents, then the plugin config can be tested
by modifying the plugin lists in tests/test.yml and
molecule/default/converge.yml.