[dpdk_telemetry] Add template for dpdk_telemetry
Config values are documented in README.rst

Support for dpdk_telemetry plugin is not yet included in the container used for
testing with molecule, so tests are not included.
Once the container supports dpdk_telemetry, then the plugin config can be tested
by modifying the plugin lists in tests/test.yml and
molecule/default/converge.yml.