Add custom types.db files
If plugins are not part of the collectd repo, (e.g. not written in C),
and new entries to types.db are needed, a custom types.db can be added
to collectd.conf.
This patch allows the config option collectd_typesdb_extra to be defiend
and passed to the role, which wil generate a new typesDB file and add it
to the main collectd.conf.