[processes] Use dict instead of list
Update processes and process_matches to use dict instead of list, which
is closer to the puppet configuration, and consistent with other plugins
such as ceph. The use of a dict to translate to the XML node in the
outputted config is more intuitive than using a list to define these
options.