[WIP][ceph] Add some logic to auto configure the ceph daemons
Currently, collectd-ceph needs the OSDs explicitly listed in order to
collectd metrics about them.

When deploying ceoph, the OSDs are distributed over multiple hosts, and
may not land in a predictable location. When using a deployment tool to
deploy both Ceph and Collectd, this becomes an issue because collectd
required the OSDs to be explicitly listed, and this might not be
possible before deployment.

This update configures collectd-ceph based on the contents of
/var/run/ceph (which is hardcoded as the OSD location previously).
If colelctd_plugin_ceph_daemons is already configured, this change has
no effect.

Potential issues: these sockets might not be there until ceph is running...
This change assumes that the OSDs are located in /var/run/ceph/ and
exist at the time of configuration.
There might be somewhere else to get the daemon list - this needs to be
checked during testing.